% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignEpitope.R
\name{assign.epitope}
\alias{assign.epitope}
\title{Assign epitopes to a genotype}
\usage{
assign.epitope(
  ind,
  dict,
  loci = c("A", "B", "C", "DRB1", "DQB1", "DQA1", "DPB1", "DPA1"),
  binary = T
)
}
\arguments{
\item{ind}{A named vector of strings with a genotype. The names of \code{ind} should be of the format "loci_d", Where "loci" is the loci of the allele and d is either "1" or "2".}
}
\value{
A list of dataframes of the dictionary.
}
\description{
This function calculates the epitype from a genotype.
}
