% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{get.aln}
\alias{get.aln}
\title{Get alignment files}
\usage{
get.aln(loci, dest = "", IMGTHLA = "Latest")
}
\arguments{
\item{loci}{A string or vector of strings of the loci to be downloaded}

\item{dest}{A string with the destination folder for the downloaded files}

\item{IMGTHLA}{A string of the IMGT/HLA version to download the alignments for}
}
\description{
This function recovers the alignment files from the IMGT/HLA github account
(https://raw.githubusercontent.com/ANHIG/IMGTHLA/). The latest version is
downloaded by default.
}
