% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{IMGTHLA.version}
\alias{IMGTHLA.version}
\title{Get IMGT/HLA version}
\usage{
IMGTHLA.version(x)
}
\arguments{
\item{x}{A string date in YYYY-MM-DD format}
}
\value{
A string with the IMGT/HLA version without field separators
}
\description{
This function takes a date and returns the IMGT/HLA version current at that
date. The function should work until 2024-01-01.
}
