% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{load.aln}
\alias{load.aln}
\title{Load the allele alignments}
\usage{
load.aln(loci, dest = "")
}
\arguments{
\item{pth}{A string of the path to the folder containing the alignment files}
}
\value{
A list of the alignments and complementary informations
}
\description{
This function loads downloaded allele alignments in .msf format.
}
