% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignEpitope.R
\name{load.dict}
\alias{load.dict}
\title{Load the eplet dictionary}
\usage{
load.dict(pth)
}
\arguments{
\item{pth}{A string of the path of the .xlsx file containing the dictionary.}
}
\value{
A list of dataframes of the dictionary.
}
\description{
This function loads a precalculated dictinary for allele to eplet conversion.
}
