% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignEpitope.R
\name{multi.assign}
\alias{multi.assign}
\title{Assign epitopes to a dataframe of genotypes}
\usage{
multi.assign(
  df,
  dict,
  loci = c("A", "B", "C", "DRB1", "DQB1", "DQA1", "DPB1", "DPA1"),
  binary = T,
  mc = F
)
}
\arguments{
\item{df}{A dataframe of the genotypes. The names of columns containing genotypes should be of the format "loci_d", Where "loci" is the loci of the allele and d is either "1" or "2".}

\item{dict}{A list of dataframes of the dictionary containing the conversion between alleles and epitype.}

\item{loci}{A string or vector of strings of the loci to be considered in the epitype.}

\item{binary}{A logical indicating whether the sum or a 0/1 value should be returned for each eplet in the epitype.}

\item{mc}{A logical indicating whether parallel processing should be used.}
}
\value{
A dataframe containing the information of \code{df} with the epytipe appended as supplementary columns.
}
\description{
This function calculates the epitype from a dataframe containing individuals as rows and genotypes in specifically named columns. Supplementary information and genotypes are kept.
}
